Fetch flyway 4.0 release

+++ flyway-core/src/main/java/org/flywaydb/core/internal/dbsupport/oracle/OracleSchema.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2010-2015 Axel Fontaine
+ * Copyright 2010-2016 Boxfuse GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -71,25 +71,35 @@ public class OracleSchema extends Schema<OracleDbSupport> {
             throw new FlywayException("Clean not supported on Oracle for user 'SYSTEM'! You should NEVER add your own objects to the SYSTEM schema!");
         }
 
-        for (String statement : generateDropStatementsForSpatialExtensions()) {
+        String user = dbSupport.doGetCurrentSchemaName();
+        boolean defaultSchemaForUser = user.equalsIgnoreCase(name);
+
+        if (!defaultSchemaForUser) {
+            LOG.warn("Cleaning schema " + name + " by a different user (" + user + "): " +
+                    "spatial extensions, queue tables, flashback tables and scheduled jobs will not be cleaned due to Oracle limitations");
+        }
+
+        for (String statement : generateDropStatementsForSpatialExtensions(defaultSchemaForUser)) {
             jdbcTemplate.execute(statement);
         }
 
-        for (String statement : generateDropStatementsForQueueTables()) {
-            try {
-                jdbcTemplate.execute(statement);
-            } catch (SQLException e) {
-                if (e.getErrorCode() == 65040) {
-                    //for dropping queue tables, a special grant is required:
-                    //GRANT EXECUTE ON DBMS_AQADM TO flyway;
-                    LOG.error("Missing required grant to clean queue tables: GRANT EXECUTE ON DBMS_AQADM");
+        if (defaultSchemaForUser) {
+            for (String statement : generateDropStatementsForQueueTables()) {
+                try {
+                    jdbcTemplate.execute(statement);
+                } catch (SQLException e) {
+                    if (e.getErrorCode() == 65040) {
+                        //for dropping queue tables, a special grant is required:
+                        //GRANT EXECUTE ON DBMS_AQADM TO flyway;
+                        LOG.error("Missing required grant to clean queue tables: GRANT EXECUTE ON DBMS_AQADM");
+                    }
+                    throw e;
                 }
-                throw e;
             }
-        }
 
-        if (flashbackAvailable()) {
-            executeAlterStatementsForFlashbackTables();
+            if (flashbackAvailable()) {
+                executeAlterStatementsForFlashbackTables();
+            }
         }
 
         for (String statement : generateDropStatementsForScheduledJobs()) {
@@ -251,10 +261,11 @@ public class OracleSchema extends Schema<OracleDbSupport> {
     /**
      * Generates the drop statements for Oracle Spatial Extensions-related database objects.
      *
+     * @param defaultSchemaForUser Whether we are currently cleaning the default schema for the logged in user.
      * @return The complete drop statements, ready to execute.
      * @throws SQLException when the drop statements could not be generated.
      */
-    private List<String> generateDropStatementsForSpatialExtensions() throws SQLException {
+    private List<String> generateDropStatementsForSpatialExtensions(boolean defaultSchemaForUser) throws SQLException {
         List<String> statements = new ArrayList<String>();
 
         if (!spatialExtensionsAvailable()) {
@@ -270,12 +281,13 @@ public class OracleSchema extends Schema<OracleDbSupport> {
             return statements;
         }
 
+        if (defaultSchemaForUser) {
+            statements.add("DELETE FROM mdsys.user_sdo_geom_metadata");
 
-        statements.add("DELETE FROM mdsys.user_sdo_geom_metadata");
-
-        List<String> indexNames = jdbcTemplate.queryForStringList("select INDEX_NAME from USER_SDO_INDEX_INFO");
-        for (String indexName : indexNames) {
-            statements.add("DROP INDEX \"" + indexName + "\"");
+            List<String> indexNames = jdbcTemplate.queryForStringList("select INDEX_NAME from USER_SDO_INDEX_INFO");
+            for (String indexName : indexNames) {
+                statements.add("DROP INDEX \"" + indexName + "\"");
+            }
         }
 
         return statements;
@@ -290,7 +302,7 @@ public class OracleSchema extends Schema<OracleDbSupport> {
     private List<String> generateDropStatementsForScheduledJobs() throws SQLException {
         List<String> statements = new ArrayList<String>();
 
-        List<String> jobNames = jdbcTemplate.queryForStringList("select JOB_NAME from USER_SCHEDULER_JOBS");
+        List<String> jobNames = jdbcTemplate.queryForStringList("select JOB_NAME from ALL_SCHEDULER_JOBS WHERE owner=?", name);
         for (String jobName : jobNames) {
             statements.add("begin DBMS_SCHEDULER.DROP_JOB(job_name => '" + jobName + "', defer => false, force => true); end;");
         }
@@ -331,29 +343,29 @@ public class OracleSchema extends Schema<OracleDbSupport> {
                 // For every table this query will count the number of references (including the transitive ones)
                 // and order the result list using that value.
                 " SELECT r FROM" +
-                "   (SELECT CONNECT_BY_ROOT t r FROM" +
-                "     (SELECT DISTINCT c1.table_name f, NVL(c2.table_name, at.table_name) t" +
-                "     FROM all_constraints c1" +
-                "       RIGHT JOIN all_constraints c2 ON c2.constraint_name = c1.r_constraint_name" +
-                "       RIGHT JOIN all_tables at ON at.table_name = c2.table_name" +
-                "     WHERE at.owner = ?" +
+                        "   (SELECT CONNECT_BY_ROOT t r FROM" +
+                        "     (SELECT DISTINCT c1.table_name f, NVL(c2.table_name, at.table_name) t" +
+                        "     FROM all_constraints c1" +
+                        "       RIGHT JOIN all_constraints c2 ON c2.constraint_name = c1.r_constraint_name" +
+                        "       RIGHT JOIN all_tables at ON at.table_name = c2.table_name" +
+                        "     WHERE at.owner = ?" +
                         // Ignore Recycle bin objects
-                "       AND at.table_name NOT LIKE 'BIN$%'" +
+                        "       AND at.table_name NOT LIKE 'BIN$%'" +
                         // Ignore Spatial Index Tables as they get dropped automatically when the index gets dropped.
-                "       AND at.table_name NOT LIKE 'MDRT_%$'" +
+                        "       AND at.table_name NOT LIKE 'MDRT_%$'" +
                         // Ignore Materialized View Logs
-                "       AND at.table_name NOT LIKE 'MLOG$%' AND at.table_name NOT LIKE 'RUPD$%'" +
+                        "       AND at.table_name NOT LIKE 'MLOG$%' AND at.table_name NOT LIKE 'RUPD$%'" +
                         // Ignore Oracle Text Index Tables
-                "       AND at.table_name NOT LIKE 'DR$%'" +
+                        "       AND at.table_name NOT LIKE 'DR$%'" +
                         // Ignore Index Organized Tables
-                "       AND at.table_name NOT LIKE 'SYS_IOT_OVER_%'" +
+                        "       AND at.table_name NOT LIKE 'SYS_IOT_OVER_%'" +
                         // Ignore Nested Tables
-                "       AND at.nested != 'YES'" +
+                        "       AND at.nested != 'YES'" +
                         // Ignore Nested Tables
-                "       AND at.secondary != 'Y')" +
-                "   CONNECT BY NOCYCLE PRIOR f = t)" +
-                " GROUP BY r" +
-                " ORDER BY COUNT(*)", name);
+                        "       AND at.secondary != 'Y')" +
+                        "   CONNECT BY NOCYCLE PRIOR f = t)" +
+                        " GROUP BY r" +
+                        " ORDER BY COUNT(*)", name);
 
         Table[] tables = new Table[tableNames.size()];
         for (int i = 0; i < tableNames.size(); i++) {
